routed the components to matching routes and restructured components
