+++ robo_advisor/src/app/app.component.html
@@ -1,10 +1,4 @@
 <header>
     <app-navbar></app-navbar>
 </header>
-
-<app-form></app-form>
 <router-outlet></router-outlet>
-
-<footer>
-    <app-footer></app-footer>
-</footer>
