+++ robo_advisor/src/app/app.component.ts
@@ -1,16 +1,21 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { FormComponent } from './components/form/form.component';
-import { LandingPageComponent } from "./landing-page/landing-page.component";
-import { NavbarComponent } from "./navbar/navbar.component";
-import { FooterComponent } from "./footer/footer.component";
+import { LandingPageComponent } from './components/landing-page/landing-page.component';
+import { NavbarComponent } from './components/navbar/navbar.component';
+import { FooterComponent } from './components/footer/footer.component';
 
 @Component({
-    selector: 'app-root',
-    standalone: true,
-    templateUrl: './app.component.html',
-    styleUrl: './app.component.css',
-    imports: [RouterOutlet, FormComponent, LandingPageComponent, NavbarComponent, FooterComponent]
+  selector: 'app-root',
+  standalone: true,
+  templateUrl: './app.component.html',
+  styleUrl: './app.component.css',
+  imports: [
+    RouterOutlet,
+    FormComponent,
+    LandingPageComponent,
+    NavbarComponent,
+    FooterComponent,
+  ],
 })
-export class AppComponent {
-}
+export class AppComponent {}
