+++ robo_advisor/src/app/app.routes.ts
@@ -1,8 +1,8 @@
 import { Routes } from '@angular/router';
-import { LandingPageComponent } from './landing-page/landing-page.component';
-import { FormsPageComponent } from './forms-page/forms-page.component';
+import { LandingPageComponent } from './components/landing-page/landing-page.component';
+import { FormComponent } from './components/form/form.component';
 
 export const routes: Routes = [
-    {path: '', component: LandingPageComponent},
-    {path: 'formspage', component: FormsPageComponent}
+  { path: '', component: LandingPageComponent },
+  { path: 'formspage',title:'user form', component: FormComponent },
 ];
