+++ robo_advisor/src/app/components/footer/footer.component.spec.ts
