+++ robo_advisor/src/app/components/landing-page/landing-page.component.spec.ts
