+++ robo_advisor/src/app/components/navbar/navbar.component.css
@@ -1,0 +1,3 @@
+.active-style {
+    color: white;
+}
