+++ robo_advisor/src/app/components/navbar/navbar.component.html
@@ -1,0 +1,41 @@
+<header class="text-gray-400 bg-[#463f3f] body-font shadow-2xl">
+  <div
+    class="container mx-auto flex flex-wrap p-5 flex-col md:flex-row items-center"
+  >
+    <a class="flex title-font font-medium items-center text-white mb-4 md:mb-0">
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        fill="none"
+        stroke="currentColor"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="2"
+        class="w-10 h-10 text-white p-2 bg-yellow-500 rounded-full"
+        viewBox="0 0 24 24"
+      >
+        <path
+          d="M12 2L2 7l10 5 10-5-10-5zM2 17l10 5 10-5M2 12l10 5 10-5"
+        ></path>
+      </svg>
+      <span class="ml-1 text-xl">Robo-App</span>
+    </a>
+    <nav
+      class="md:ml-auto flex flex-wrap items-center text-base justify-center"
+    >
+      <a
+        class="mr-5 hover:text-white text-lg"
+        [routerLink]="['']"
+        routerLinkActive="active  active-style"
+        [routerLinkActiveOptions]="{ exact: true }"
+        >Home</a
+      >
+      <a
+        class="mr-5 hover:text-white text-lg"
+        [routerLink]="['formspage']"
+        routerLinkActive="active active-style"
+        [routerLinkActiveOptions]="{ exact: true }"
+        >GET STARTED</a
+      >
+    </nav>
+  </div>
+</header>
