+++ robo_advisor/src/app/components/navbar/navbar.component.spec.ts
