--- robo_advisor/src/app/forms-page/forms-page.component.html
@@ -1,3 +1,0 @@
-<h1>{{ title }}</h1>
-<p>This is the forms page!<br>
-    Would have loved to add some stuffs here but I'm waiting for IG.</p>
