--- robo_advisor/src/app/forms-page/forms-page.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { FormsPageComponent } from './forms-page.component';
-
-describe('FormsPageComponent', () => {
-  let component: FormsPageComponent;
-  let fixture: ComponentFixture<FormsPageComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [FormsPageComponent]
-    })
-    .compileComponents();
-    
-    fixture = TestBed.createComponent(FormsPageComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
