--- robo_advisor/src/app/forms-page/forms-page.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-forms-page',
-  standalone: true,
-  imports: [],
-  templateUrl: './forms-page.component.html',
-  styleUrl: './forms-page.component.css'
-})
-export class FormsPageComponent {
-  title = 'Forms page';
-}
