--- robo_advisor/src/app/landing-page/landing-page.component.ts
@@ -1,15 +1,0 @@
-import { Component } from '@angular/core';
-import { FooterComponent } from "../footer/footer.component";
-import { NavbarComponent } from "../navbar/navbar.component";
-import { RouterLink, RouterLinkActive} from '@angular/router';
-
-@Component({
-    selector: 'app-landing-page',
-    standalone: true,
-    templateUrl: './landing-page.component.html',
-    styleUrl: './landing-page.component.css',
-    imports: [FooterComponent, NavbarComponent, RouterLink, RouterLinkActive]
-})
-export class LandingPageComponent {
-  title = 'Landing Page works!';
-}
