--- robo_advisor/src/app/navbar/navbar.component.html
@@ -1,3 +1,0 @@
-<h4>
-    <i>That's a navbar right there!</i>
-</h4>
